Create sample_c_sh_method
Here are the steps to create a POST object and build a signature to the Rapyd API spec:

Start by creating a dictionary object that contains the parameters for the API request. This dictionary should include the required and optional parameters for the API endpoint you are trying to access.

Serialize the dictionary object to a JSON string.

Generate a Unix timestamp in seconds and add it to the dictionary object. This timestamp should be included in the headers of your API request.

Generate a random nonce and add it to the dictionary object. This nonce should also be included in the headers of your API request.

Build the signature for the API request by concatenating the HTTP method, endpoint URL, timestamp, nonce, and JSON string.

Hash the concatenated string with HMAC-SHA256 using your API secret key as the key.

Encode the hashed signature as Base64.

Include the signature and other required headers in your API request.

Post the API request to the appropriate Rapyd endpoint.